requirements.txt
Django==2.2.7
pytz==2019.3
sqlparse==0.3.0
django-cors-headers==2.2.0
djangorestframework==3.9.3
postgres==3.0.0
psycopg2-binary==2.8.4
psycopg2-pool==1.1